+++ blockchain/message_pool/src/msgpool/mod.rs
@@ -325,15 +325,11 @@ pub(crate) fn add_to_selected_msgs(
 
 #[cfg(test)]
 pub mod tests {
-    #[cfg(feature = "slow_tests")]
-    use std::borrow::BorrowMut;
-    #[cfg(feature = "slow_tests")]
-    use std::time::Duration;
+    use std::{borrow::BorrowMut, time::Duration};
 
     use forest_blocks::Tipset;
     use forest_key_management::{KeyStore, KeyStoreConfig, Wallet};
     use forest_message::SignedMessage;
-    #[cfg(feature = "slow_tests")]
     use forest_networks::ChainConfig;
     use forest_shim::{
         address::Address,
@@ -341,15 +337,15 @@ pub mod tests {
         econ::TokenAmount,
         message::{Message, Message_v3},
     };
-    #[cfg(feature = "slow_tests")]
     use num_traits::Zero;
     use test_provider::*;
     use tokio::task::JoinSet;
 
     use super::*;
-    #[cfg(feature = "slow_tests")]
-    use crate::msg_chain::{create_message_chains, Chains};
-    use crate::msg_pool::MessagePool;
+    use crate::{
+        msg_chain::{create_message_chains, Chains},
+        msg_pool::MessagePool,
+    };
 
     pub fn create_smsg(
         to: &Address,
@@ -371,7 +367,35 @@ pub mod tests {
         .into();
         let msg_signing_bytes = umsg.cid().unwrap().to_bytes();
         let sig = wallet.sign(from, msg_signing_bytes.as_slice()).unwrap();
-        SignedMessage::new_from_parts(umsg, sig).unwrap()
+        SignedMessage::new_unchecked(umsg, sig)
+    }
+
+    // Create a fake signed message with a dummy signature. While the signature is
+    // not valid, it has been added to the validation cache and the message will
+    // appear authentic.
+    pub fn create_fake_smsg(
+        pool: &MessagePool<TestApi>,
+        to: &Address,
+        from: &Address,
+        sequence: u64,
+        gas_limit: i64,
+        gas_price: u64,
+    ) -> SignedMessage {
+        let umsg: Message = Message_v3 {
+            to: to.into(),
+            from: from.into(),
+            sequence,
+            gas_limit: gas_limit as u64,
+            gas_fee_cap: TokenAmount::from_atto(gas_price + 100).into(),
+            gas_premium: TokenAmount::from_atto(gas_price).into(),
+            ..Message_v3::default()
+        }
+        .into();
+        let sig = Signature::new_secp256k1(vec![]);
+        let signed = SignedMessage::new_unchecked(umsg, sig);
+        let cid = signed.cid().unwrap();
+        pool.sig_val_cache.lock().put(cid, ());
+        signed
     }
 
     #[tokio::test]
@@ -542,7 +566,6 @@ pub mod tests {
         assert_eq!(p.len(), 3);
     }
 
-    #[cfg(feature = "slow_tests")]
     #[tokio::test]
     async fn test_async_message_pool() {
         let keystore = KeyStore::new(KeyStoreConfig::Memory).unwrap();
@@ -591,7 +614,6 @@ pub mod tests {
         assert_eq!(cur_ts.as_ref(), &tipset);
     }
 
-    #[cfg(feature = "slow_tests")]
     #[tokio::test]
     async fn test_msg_chains() {
         let keystore = KeyStore::new(KeyStoreConfig::Memory).unwrap();